add demapper_test projects for VS2010
git-svn-id: svn+ssh://svn.code.sf.net/p/itpp/svn/itpp/trunk@1882 f17a810d-92cc-4fca-aeb8-6e32867107aa

+++ tests/demapper_test.cpp
@@ -62,7 +62,7 @@ int main()
     bvec bits = randb(perm_len);
 
     //QAM modulation
-    cvec em = mod.modulate_bits(bits)/sqrt(em_antennas);//normalize emitted symbols
+    cvec em = mod.modulate_bits(bits)/sqrt(double(em_antennas));//normalize emitted symbols
 
     //ST code
     cmat S = st_block_code.encode(em);
